create a detail activitiy and make it pop up
Detail activity pops up on click of an item from the weatherfragment
The clickingmakes an intent with an extra text. The detail activity will then get the intent and if its not null and it has extra text - it will get the view by id and set the text to be the forecast.